Setting the correct type for notes onDelete() ans adding onUpdate()
The _noteCard() in example_full\lib\pages\edit_contact_page.dart file erroneously points to _contact.groups for onDelete(). This throws an null/empty error when one adds a note and tries to delete it (as is). This fix ensures the notes property is the one targeted. Adding onUpdate() that was initially null