Setting correct PhoneLabel type for PhoneLabel.mobile
Both iOS and Android support 'mobile' PhoneLabel. The current parser assigns type('cell') + ('text') or ('msg') which then miss-assigns the label when the contact is read or scanned. The current setup passes TEL;TYPE=cell,msg which is a custom label and not 'mobile' as intended. This PR fixes this by exclusively returning types.add('mobile') for PhoneLabel.mobile